+++ docs/changelog.md
@@ -1,6 +1,6 @@
 # Change log
 
-## VERSION
+## 2210.645
 
 * Adds `OnlyOfficeDesktopEditors`, `OperaCryptoBrowser`. `AdobeConnect` [#382](https://github.com/aaronparker/evergreen/issues/382)
 * Updates channels for `OperaBrowser` [#299](https://github.com/aaronparker/evergreen/issues/299)